usrp2: Include variable bit gain set by register after CIC filter.
Decimation by N creates a possible log2(N)/2 bits of additional
precision. Currently, this extra precision is lost in the DDC chain by
internal conversions down to 17-18 bits from 24 bits for multiplication.
This patch inserts a variable bit gain after the CIC filter and before
any of the decimation precision is lost, allowing that precision to be
kept at the expense of higher order bits. When the signal levels are low
enough so that the higher order bits are not used, this is a great
tradeoff to make.

Using this requires a change to the UHD host code to specify the bit
gain.

This commit adds variable bitgain with clipping as a separate module
(variable_part_select_and_clip) that is called from the cic decimator.
The cic_dec_shifter module was modified so that a multiplexer is
inferred, reducing its footprint. The cic decimator pipeline was also
fixed so that the strobe matches the earliest valid output, and an
additional pipeline stage was added for the variable bitgain in order to
meet timing requirements.